Share i18n sources
