+++ electron/tsconfig.json
@@ -11,5 +11,5 @@
     "outDir": "../public/",
     "pretty": true
   },
-  "include": ["./"]
+  "include": ["./", "../src/shared"]
 }
