+++ src/shared/i18n/types.ts
@@ -1,0 +1,5 @@
+export enum Locale {
+  EN = 'en',
+  FR = 'fr',
+  DE = 'de'
+}
