+++ src/shared/ipc/messages.ts
@@ -1,0 +1,5 @@
+enum IPCMessages {
+  UPDATE_LANG = 'UPDATE_LANG'
+}
+
+export default IPCMessages
